Renamed all components
