+++ app/layout.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import './globals.css'
-import Navbar from '@/components/ui/navbar'
+import Navbar from '@/components/ui/Navbar'
 
-import { ThemeProvider } from "@/components/theme-provider"
+import { ThemeProvider } from "@/components/ThemeProvider"
 import { cn } from '@/lib/utils'
 import AuthProvider from './Provider'
