--- app/posts/page.tsx
@@ -1,8 +1,0 @@
-
-export default function Posts() {
-    return (
-        <>
-            Posts go here
-        </>
-    )
-}
