+++ components/Posts.tsx
@@ -1,0 +1,19 @@
+
+export default async function Posts() {
+
+    async function getPosts() {
+
+
+    }
+
+    const data = await getPosts();
+
+    return (
+
+        <section id="posts" className="flex flex-col gap-2 justify-center">
+            <div className="">
+                Post 1
+            </div>
+        </section>
+    )
+}
