+++ components/ui/toaster.tsx
@@ -7,8 +7,8 @@ import {
   ToastProvider,
   ToastTitle,
   ToastViewport,
-} from "@/components/ui/toast"
-import { useToast } from "@/components/ui/use-toast"
+} from "@/components/ui/Toast"
+import { useToast } from "@/components/ui/UseToast"
 
 export function Toaster() {
   const { toasts } = useToast()
